+++ packages/core/protocols/src/index.ts
@@ -13,3 +13,4 @@ export * from './automerge.js';
 export * from './edge.js';
 export * from './queue.js';
 export * from './space-archive.js';
+export * from './edge-error.js';
